% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{calculate_positions}
\alias{calculate_positions}
\title{Calculate tube coordinates and angles}
\usage{
calculate_positions(N)
}
\arguments{
\item{N}{Scalar integer, the number of total positions in the rotor.}
}
\value{
Data.frame with one row for each element of \code{r} and columns
\itemize{
\item n: position, as provided by the user
\item theta: angle
\item x: x-coordinate
\item y: y-coordinate
}
}
\description{
Calculate tube coordinates and angles
}
\examples{
calculate_positions(32)
}
