% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{calculate_tube_coordinate}
\alias{calculate_tube_coordinate}
\title{Calculate the x and y coordinates for a tube based on its angle}
\usage{
calculate_tube_coordinate(theta)
}
\arguments{
\item{theta}{Numeric vector, the rotational angle(s) (theta)}
}
\value{
Data.frame with one row for each element of \code{r} and columns
\itemize{
\item x: x-coordinate
\item y: y-coordinate
}
}
\description{
Calculate the x and y coordinates for a tube based on its angle
}
\examples{
theta <- calculate_tube_angle(32)
calculate_tube_coordinate(theta)
}
