% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{compute_S2SI_BigRaster}
\alias{compute_S2SI_BigRaster}
\title{this function aims at computing spectral indices from big raster data
it computes the spectral indices based on clsest bands with Sentinel-2
and assumes that the bands of the S2 data follow this order
wavelength	= {496.6, 560.0, 664.5, 703.9, 740.2, 782.5, 835.1, 864.8, 1613.7, 2202.4}
Full description of the indices can be found here:
https://www.sentinel-hub.com/eotaxonomy/indices}
\usage{
compute_S2SI_BigRaster(
  raster_path,
  PathOut,
  MaskRaster = FALSE,
  SensorBands,
  Sel_Indices = "ALL",
  ReflFactor = 1,
  Offset = 0,
  S2Bands = data.frame(B2 = 492.7, B3 = 559.8, B4 = 664.6, B5 = 704.1, B6 = 740.5, B7 =
    782.8, B8 = 832.8, B8A = 864.7, B11 = 1613.7, B12 = 2202.4)
)
}
\arguments{
\item{raster_path}{character. path for the big raster}

\item{PathOut}{character. path for output directory}

\item{MaskRaster}{character. path for mask file corresponding to raster_path}

\item{SensorBands}{numeric. wavelength in nanometers of the spectral bands of raster_path}

\item{Sel_Indices}{list. list of spectral indices to be computed}

\item{ReflFactor}{numeric. multiplying factor used to write reflectance in image (==10000 for S2)}

\item{Offset}{numeric. offset (when Refl between 0 and 1) to be applied on reflectance. Useful to avoid zero values}

\item{S2Bands}{numeric. wavelength of the spectral bands corresponding to S2 (default = S2A)}
}
\value{
list. list of rasters corresponding to spectral indices
}
\description{
this function aims at computing spectral indices from big raster data
it computes the spectral indices based on clsest bands with Sentinel-2
and assumes that the bands of the S2 data follow this order
wavelength	= {496.6, 560.0, 664.5, 703.9, 740.2, 782.5, 835.1, 864.8, 1613.7, 2202.4}
Full description of the indices can be found here:
https://www.sentinel-hub.com/eotaxonomy/indices
}
