% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{compute_S2SI_from_Sensor}
\alias{compute_S2SI_from_Sensor}
\title{this function aims at computing spectral indices from Sensor reflectance data.
it computes the spectral indices based on their computation with Sentinel-2
and assumes that the bands of the S2 data follow this order
wavelength	= {496.6, 560.0, 664.5, 703.9, 740.2, 782.5, 835.1, 864.8, 1613.7, 2202.4}
Full description of the indices can be found here:
https://www.sentinel-hub.com/eotaxonomy/indices}
\usage{
compute_S2SI_from_Sensor(
  Refl,
  SensorBands,
  Sel_Indices = "ALL",
  nbCPU = 1,
  S2Bands = data.frame(B2 = 492.7, B3 = 559.8, B4 = 664.6, B5 = 704.1, B6 = 740.5, B7 =
    782.8, B8 = 832.8, B8A = 864.7, B11 = 1613.7, B12 = 2202.4)
)
}
\arguments{
\item{Refl}{numeric. Reflectance dataset defined in matrix}

\item{SensorBands}{numeric. wavelength of the spectral bands corresponding to reflectance}

\item{Sel_Indices}{list. list of spectral indices to be computed}

\item{nbCPU}{numeric. number of CPU if parallel processing required}

\item{S2Bands}{numeric. wavelength of the spectral bands corresponding to S2 (default = S2A)}
}
\value{
list. includes
\itemize{
\item SpectralIndices: List of spectral indices computed from the reflectance initially provided
\item listIndices: list of spectral indices computable with the function
}
}
\description{
this function aims at computing spectral indices from Sensor reflectance data.
it computes the spectral indices based on their computation with Sentinel-2
and assumes that the bands of the S2 data follow this order
wavelength	= {496.6, 560.0, 664.5, 703.9, 740.2, 782.5, 835.1, 864.8, 1613.7, 2202.4}
Full description of the indices can be found here:
https://www.sentinel-hub.com/eotaxonomy/indices
}
