% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{Compute_SI_CR_Corr}
\alias{Compute_SI_CR_Corr}
\title{This function computes the correlation between a set of vegetation properties
and a continuum removed spectral index}
\usage{
Compute_SI_CR_Corr(listBands, Refl, SensorBands, BPvars)
}
\arguments{
\item{listBands}{numeric. bands used in the expression}

\item{Refl}{numeric. reflectance matrix with ncol = nbands and nrows = nsamples}

\item{SensorBands}{numeric. vector containing central wavelength for each spectral band of ReflMat}

\item{BPvars}{numeric. Biophysical properties as matrix with ncol = nBPvars and nrows = nsamples}
}
\value{
list. includes band Combinations and corresponding correlation between biophysical variables and SI
}
\description{
This function computes the correlation between a set of vegetation properties
and a continuum removed spectral index
}
