% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{Compute_SI_fromExp_Corr}
\alias{Compute_SI_fromExp_Corr}
\title{This function computes the correlation between a set of vegetation properties
and a spectral index}
\usage{
Compute_SI_fromExp_Corr(
  listBands,
  Refl,
  SensorBands,
  ExpressIndex,
  BPvars,
  ReflFactor = 1,
  NameIndex = "SI"
)
}
\arguments{
\item{listBands}{numeric. bands used in the expression}

\item{Refl}{numeric. reflectance matrix with ncol = nbands and nrows = nsamples}

\item{SensorBands}{numeric. vector containing central wavelength for each spectral band of ReflMat}

\item{ExpressIndex}{character. expression corresponding to the spectral index to be explored ()}

\item{BPvars}{numeric. Biophysical properties as matrix with ncol = nBPvars and nrows = nsamples}

\item{ReflFactor}{numeric. multplying factor for reflectance}

\item{NameIndex}{character. name of spectral index}
}
\value{
list. includes band Combinations and corresponding correlation between biophysical variables and SI
}
\description{
This function computes the correlation between a set of vegetation properties
and a spectral index
}
