% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{is_balanced}
\alias{is_balanced}
\title{Determine whether a rotor is balanced}
\usage{
is_balanced(df, tolerance = 0.001)
}
\arguments{
\item{df}{Data.frame with columns
\itemize{
\item x: x-coordinate (numeric)
\item y: y-coordinate (numeric)
\item occupied: boolean
}}
}
\value{
Scalar boolean
df <- calculate_positions(30)
df$occupied <- FALSE
df\link{c(1, 11, 14, 15, 21, 29, 30), "occupied"} <- TRUE
is_balanced(df)  # TRUE
df\link{29, "occupied"} <- TRUE
is_balanced(df)  # FALSE
}
\description{
Determine whether a rotor is balanced
}
