% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{listIndices_spinR}
\alias{listIndices_spinR}
\title{This function lists the spectral indices defined in spinR}
\usage{
listIndices_spinR()
}
\value{
list.
}
\description{
This function lists the spectral indices defined in spinR
}
