% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{Optimal_SI_CR}
\alias{Optimal_SI_CR}
\title{This function computes the correlation between a set of vegetation properties
and all combinations of spectral bands corresponding to a given type of spectral index}
\usage{
Optimal_SI_CR(ReflMat, Spectral_Bands, BPvars, repeats.allowed = FALSE)
}
\arguments{
\item{ReflMat}{numeric. reflectance matrix with ncol = nbands and nrows = nsamples}

\item{Spectral_Bands}{numeric. vector containing central wavelength for each spectral band of ReflMat}

\item{BPvars}{numeric. Biophysical properties as matrix with ncol = nBPvars and nrows = nsamples}

\item{repeats.allowed}{boolean. should combinations include repetitions? set false if SI expression is symmetrical to gain time}
}
\value{
list. includes band Combinations and corresponding correlation between biophysical variables and SI
}
\description{
This function computes the correlation between a set of vegetation properties
and all combinations of spectral bands corresponding to a given type of spectral index
}
