% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rotor}
\alias{plot_rotor}
\title{Plot rotor occupancy}
\usage{
plot_rotor(df, radius = 1, size = 10, text_size = 5)
}
\arguments{
\item{df}{Data.frame with columns
\itemize{
\item x: x-coordinate (numeric)
\item y: y-coordinate (numeric)
\item occupied (optional): boolean
if the \code{occupied} column is missing, all positions will be considered empty.
}}

\item{size}{Scalar numeric, symbol size}

\item{text_size}{Scalar numeric, text size}
}
\description{
Plot rotor occupancy
}
\examples{
df <- calculate_positions(30)
df$occupied <- FALSE
df[c(1, 11, 14, 15, 21, 29, 30), "occupied"] <- TRUE
plot_rotor(df)
}
