% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{spectralindices_from_raster}
\alias{spectralindices_from_raster}
\title{This function computes spectral indices from raster data}
\usage{
spectralindices_from_raster(
  input_raster_path,
  input_rast_wl,
  output_dir,
  output_rasters = NULL,
  SI_list,
  input_mask_path = NULL,
  ReflFactor = 10000,
  nbCPU = 1,
  filetype = "GTiff",
  overwrite = TRUE,
  SIFactor = 1,
  datatype = "FLT4S"
)
}
\arguments{
\item{input_raster_path}{character. raster path / list of raster paths}

\item{input_rast_wl}{numeric. vector containing central wavelength for each spectral band in the image}

\item{output_dir}{character. main directory where SI will be written}

\item{output_rasters}{character. list of paths corresponding to each path}

\item{SI_list}{character. spectral indices to be computed (cf package spinR)}

\item{input_mask_path}{character. path for mask}

\item{ReflFactor}{numeric. multiplying factor used to write reflectance in image (==10000 for S2)}

\item{nbCPU}{numeric. number of CPU for parallel processing}

\item{filetype}{character. driver}

\item{overwrite}{boolean. set TRUE to overwrite existing files}

\item{SIFactor}{numeric. multiplying factor to apply to spectral index}

\item{datatype}{character. data type to save spectral index with terra. Default = 'FLT4S'}
}
\value{
output_rasters list. list of path for SI rasters
}
\description{
This function computes spectral indices from raster data
}
