% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{spinrApp}
\alias{spinrApp}
\title{Start the spinr shiny application}
\usage{
spinrApp(appDir = system.file("shiny", package = "spinr"), ...)
}
\arguments{
\item{appDir}{Scalar character, the path to the \code{spinr} shiny app}

\item{...}{Additional arguments passed on to the \code{\link[shiny:runApp]{shiny::runApp()}} function}
}
\value{
Nothing, the shiny app will be started in a separate window
}
\description{
Start the spinr shiny application
}
\examples{
\dontrun{
spinrApp()
}
}
