% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_BigRaster.R
\name{stack_BigRaster}
\alias{stack_BigRaster}
\title{this function aims at stacking individual rasters into a unique raster file
when the raster size is important.
the file is then written in ENVI BIL as it is written line per line}
\usage{
stack_BigRaster(
  list_rasters,
  stack_file,
  MaskRaster = FALSE,
  names_rasters = FALSE
)
}
\arguments{
\item{list_rasters}{list. list of raster files to be stacked}

\item{stack_file}{character. path for output file to be stacked}

\item{MaskRaster}{character. path for mask file corresponding to raster_path}

\item{names_rasters}{character. name for each of the rasters in list_rasters, if not already named}
}
\value{
list. list of rasters corresponding to spectral indices
}
\description{
this function aims at stacking individual rasters into a unique raster file
when the raster size is important.
the file is then written in ENVI BIL as it is written line per line
}
