% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{sub_SI_CR_Corr}
\alias{sub_SI_CR_Corr}
\title{This function computes the correlation between a set of vegetation properties
and a set of combinations of spectral bands corresponding to a given type of spectral index}
\usage{
sub_SI_CR_Corr(Bands, Refl, SensorBands, BPvars, p = NULL)
}
\arguments{
\item{Bands}{numeric.}

\item{Refl}{numeric. reflectance matrix with ncol = nbands and nrows = nsamples}

\item{SensorBands}{numeric. vector containing central wavelength for each spectral band of ReflMat}

\item{BPvars}{numeric. Biophysical properties as matrix with ncol = nBPvars and nrows = nsamples}

\item{p}{function.}
}
\value{
SIopt list. includes band Combinations and corresponding correlation between biophysical variables and SI
}
\description{
This function computes the correlation between a set of vegetation properties
and a set of combinations of spectral bands corresponding to a given type of spectral index
}
