% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{wrapperBig_SI_S2perband}
\alias{wrapperBig_SI_S2perband}
\title{this function aims at computing spectral indices from a list of rasters
with the function apply_bigRaster from the package bigRaster
(https://gitlab.com/jbferet/bigRaster)
The list of rasters corresponds to individual bands.
It computes the spectral indices based on closest bands with Sentinel-2
and assumes that the bands of the S2 data follow this order
wavelength	= {496.6, 560.0, 664.5, 703.9, 740.2, 782.5, 835.1, 864.8, 1613.7, 2202.4}
Full description of the indices can be found here:
https://www.sentinel-hub.com/eotaxonomy/indices
or by calling spinR::listIndices_spinR()}
\usage{
wrapperBig_SI_S2perband(input_data, input_args)
}
\arguments{
\item{input_data}{list. Image data chunk and corresponding mask if available}

\item{input_args}{list. arguments used to compute spectral indices
\itemize{
\item input_args$SensorBands
\item input_args$S2Bands
\item input_args$ReflFactor
\item input_args$SI
\item input_args$nbCPU
}}
}
\value{
list. list of rasters corresponding to spectral indices
}
\description{
this function aims at computing spectral indices from a list of rasters
with the function apply_bigRaster from the package bigRaster
(https://gitlab.com/jbferet/bigRaster)
The list of rasters corresponds to individual bands.
It computes the spectral indices based on closest bands with Sentinel-2
and assumes that the bands of the S2 data follow this order
wavelength	= {496.6, 560.0, 664.5, 703.9, 740.2, 782.5, 835.1, 864.8, 1613.7, 2202.4}
Full description of the indices can be found here:
https://www.sentinel-hub.com/eotaxonomy/indices
or by calling spinR::listIndices_spinR()
}
